LayoutTests/accessibility/img-alt-tag-only-whitespace.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div tabindex="0" role="group" id="images">
<img id="firstimage" alt=" " src="resources/cake.png"><br>
<img alt="	" src="resources/cake.png"><br>
<img alt="   Image " src="resources/cake.png"><br>
</div>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that images with alt tags that only have white space are ignored.");

    if (window.accessibilityController) {

        document.getElementById("images").focus();
        var imagesGroup = accessibilityController.focusedElement;
        shouldBe("imagesGroup.childrenCount", "1");
        if (accessibilityController.platformName != "atk")
            console.log("Image description: " + imagesGroup.childAtIndex(0).description);
        else
            console.log("Image title: " + imagesGroup.childAtIndex(0).title);
    }

</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
